Let’s make this more correct.
size_t will vary depending on implementation so it’s better to make sure this won’t become a problem later on.

+++ game/screen_sing.cc
@@ -609,7 +609,7 @@ void ScreenSing::drawMenu() {
 		txt->dimensions.middle(x).center(y);
 		txt->draw(it->getName());
 		if (it->value == &m_vocalTracks[player]) {
-					if(boost::lexical_cast<unsigned long>(player) < analyzers.size()) {
+					if(boost::lexical_cast<size_t>(player) < analyzers.size()) {
 				Color color = MicrophoneColor::get(analyzers[player].getId());
 				ColorTrans c(color);
 				m_player_icon->dimensions.right(x).fixedHeight(0.040).center(y);
